editoast: move infra errors to generated_data
