+++ editoast/src/models/infra.rs
@@ -1,4 +1,3 @@
-pub mod errors;
 mod object_queryable;
 mod railjson_data;
 mod route_from_waypoint_result;
